\docType{data}
\name{geneAnnot}
\alias{geneAnnot}
\title{CDKN2A gene locus annotation}
\description{
  A dataset containing the annotations for the CDKN2A 
  locus.
}
\usage{data(geneAnnot)}
\format{A \code{GRanges} object}
\source{
  The Cancer Genome Atlas Research Network. (2012)
  Comprehensive genomic characterization of squamous cell
  lung cancers. Nature 489: 519-525.
}
\keyword{datasets}

