\docType{data}
\name{geneDepth}
\alias{geneDepth}
\title{Coverage matrix across CDKN2A gene locus}
\description{
  A dataset containing read depths for 179 lung squamous cell carcinoma
  samples across the CDKN2A locus.
}
\usage{data(geneDepth)}
\format{A \eqn{2078 \times 179}{2078 x 179} \code{data.frame} of read depth (coverage). Each 
  column corresponds to a sample and each row to a base position along the
  CDKN2A locus.  These RNA-Seq read counts are a subset from 179 lung squamous 
  cell tumor samples sequenced as part of the Cancer Genome Atlas.}
\source{
  The Cancer Genome Atlas Research Network. (2012)
  Comprehensive genomic characterization of squamous cell
  lung cancers. Nature 489: 519-525.
}
\keyword{datasets}

