\docType{package}
\name{SigFuge-package}
\alias{SigFuge}
\alias{SigFuge-package}
\title{SigFuge}
\description{
  Tests significance of clustering in RNA-seq data.
}
\details{
  \code{\link{SFpval}} computes a \eqn{p}-value for
  significance of clustering for RNA-seq data, and
  \code{\link{SFfigure}} produces accompanying figures.
}
\author{
  Patrick Kimes \email{pkimes@live.unc.edu}
}
\keyword{package}

